refactor: zero lint errors across monorepo (wave 3 of CI hardening)
Cleanup pass — fixes the final 52 lint errors spread across ~40 files
after waves 1 and 2. Monorepo now exits 0 on `npm run lint` (17
warnings remain, all react-refresh/exhaustive-deps, acceptable).

## Breakdown (47 errors eliminated this pass)
- 22 no-misused-promises: wrapped async handlers as
  `onClick={() => { void handleFoo(); }}` or equivalent. Affected ~20
  admin pages (AddApplication, ViewApplication, CentrePayments,
  AddCohort, AddCourse, AddLesson, AddEntranceExam, AddExam,
  ExamEvaluation, PaymentStatus, Instructors, ProgramCourses,
  4 Settings pages, Students, TrainingVideos, AdminUsers, SuperAdmin).
- 18 no-unused-vars: removed unused imports (Button, CardHeader, X,
  asNumber, asString, formatDate, toRecords) or renamed to `_name`
  (onNavigate, onReload, batches). Deleted 3 unused helper functions
  in the API (idString, two copies of toInteger).
- 7 no-base-to-string: replaced `String(v)` with `asString(v)` from
  admin-data-utils.ts on unknown-typed fields (ViewSubmissions,
  AssociateTarget, CounsellorTarget, Certificates, ProgramCourses).
- 3 no-floating-promises: added `void` to fire-and-forget calls
  (deleteEntranceExam, deleteQuestion, loadCourses().then).
- 1 no-unsafe-argument: `Buffer.from(chunk as Uint8Array)` in
  routes/content.ts.
- 1 no-redundant-type-constituents: `course_id: string | ''` →
  `course_id: string` in auth/types.ts.

## Final state across waves 1-3
- Lint: 507 → 0 errors (17 warnings remain, mostly react-refresh).
- Typecheck: 0 errors across api + web + frontend-core.
- Build: all workspaces pass.
- Zero `// eslint-disable-next-line` comments added across all 3 waves.
- Zero behavior changes.

+++ apps/api/src/routes/content.ts
@@ -1145,7 +1145,7 @@ export function registerContentRoutes(
       const key = `uploads/${Date.now()}-${Math.random().toString(36).slice(2)}.${ext}`;
       const chunks: Buffer[] = [];
       for await (const chunk of file.file) {
-        chunks.push(Buffer.from(chunk));
+        chunks.push(Buffer.from(chunk as Uint8Array));
       }
       const body = Buffer.concat(chunks);
       const result = await storage.uploadObject({
